+++ test/ops/test_aggregate.py
@@ -7,9 +7,14 @@ import numpy as np
 import pandas as pd
 import xarray as xr
 
+from xcube.core.new import new_cube
+
 from esa_climate_toolbox.core.op import OP_REGISTRY
 from esa_climate_toolbox.ops import adjust_temporal_attrs
-from esa_climate_toolbox.ops import climatology, temporal_aggregation, reduce
+from esa_climate_toolbox.ops import climatology
+from esa_climate_toolbox.ops import reduce
+from esa_climate_toolbox.ops import statistics
+from esa_climate_toolbox.ops import temporal_aggregation
 from esa_climate_toolbox.util.misc import object_to_qualified_name
 from esa_climate_toolbox.util.monitor import ConsoleMonitor
 
@@ -392,3 +397,55 @@ class TestReduce(TestCase):
             'time': pd.date_range('2000-01-01', '2000-12-31')})
 
         self.assertTrue(actual.broadcast_equals(ex))
+
+
+class TestStatistics(TestCase):
+
+    def test_statistics_mean(self):
+        def sst_func(t, y, x):
+            return t * 2.0 + x * 0.1
+
+        def chl_func(t, y, x):
+            return y * 1.0 + x * 0.1
+
+        def aot_func(t, y, x):
+            return y * 0.1 + x * 1.0
+
+        cube = new_cube(
+            width=16, height=8, variables=dict(sst=sst_func, chl=chl_func, aot=aot_func)
+        )
+
+        res = statistics(
+            cube, var=["chl", "aot"],
+            methods=["count", "mean", "median", "sum", "std", "min", "max"]
+        )
+
+        self.assertIsNotNone(res)
+        self.assertIn("chl_count", res.data_vars)
+        self.assertEqual(635, res["chl_count"].values)
+        self.assertIn("chl_mean", res.data_vars)
+        self.assertAlmostEqual(4.25, res["chl_mean"].values, 6)
+        self.assertIn("chl_median", res.data_vars)
+        self.assertAlmostEqual(4.25, res["chl_median"].values, 6)
+        self.assertIn("chl_sum", res.data_vars)
+        self.assertAlmostEqual(2720, res["chl_sum"].values, 6)
+        self.assertIn("chl_std", res.data_vars)
+        self.assertAlmostEqual(2.337199, res["chl_std"].values, 6)
+        self.assertIn("chl_min", res.data_vars)
+        self.assertAlmostEqual(0, res["chl_min"].values, 6)
+        self.assertIn("chl_max", res.data_vars)
+        self.assertAlmostEqual(8.5, res["chl_max"].values, 8)
+        self.assertIn("aot_count", res.data_vars)
+        self.assertEqual(635, res["aot_count"].values, 6)
+        self.assertIn("aot_mean", res.data_vars)
+        self.assertAlmostEqual(7.85, res["aot_mean"].values, 6)
+        self.assertIn("aot_median", res.data_vars)
+        self.assertAlmostEqual(7.85, res["aot_median"].values, 6)
+        self.assertIn("aot_sum", res.data_vars)
+        self.assertAlmostEqual(5024, res["aot_sum"].values, 6)
+        self.assertIn("aot_std", res.data_vars)
+        self.assertAlmostEqual(4.615463, res["aot_std"].values, 6)
+        self.assertIn("aot_min", res.data_vars)
+        self.assertAlmostEqual(0, res["aot_min"].values, 6)
+        self.assertIn("aot_max", res.data_vars)
+        self.assertAlmostEqual(15.7, res["aot_max"].values, 6)